Ported HLU fixes: unhandled keyvalues, type correctness (iStringNull, HSPRITE, int->float), sscanf/fread/getcwd error checks, buffer sizes, format strings, zombie pitch var

+++ dlls/h_cycler.cpp
@@ -400,7 +400,7 @@ class CWreckage : public CBaseMonster
 	void Precache( void );
 	void Think( void );
 
-	int m_flStartTime;
+	float m_flStartTime;
 };
 /*TYPEDESCRIPTION	CWreckage::m_SaveData[] = 
 {
